Added a simple script for setting up the CouchDB
Script can create a databse in CouchDB for crafts and also insert an
initial configuration and add attachments to it. It can also delete this
database. This script will be used later to insert all of the views. It
is also the first piece of code to use the new CouchDB driver.