Changed the way couch functions are added and added a list function.
Couch functions are now all in the couch directory since there are now
two types of functions being imported. Also added an "update" command so
that the views and lists can be updated without resetting the database.
Added a list function to return historical data in a format which can be
used by the web interface.