Major overhaul of the metrics and views.
In order to maintain a more uniform data representation, some major
changes have been made to the way metrics are stored and presented.

Collections are now limited to containing data from a single role. This
makes the data much easier to handle and is not an unreasonable
contraint.

All views are now keyed on [role, timestamp] so that views can be easily
filtered by role and allows for easy handling of multiple clients in a
single db. Additionally, each document type now has its own single view.

Aggregation is now done in a view to guarantee consistency.

Predictions are stored in their documents at the same place in the
hierarchy as the would be in the aggregate view. This allows for
predictions to be done on multiple metrics while still preserving
others.